requirements.txt
fastapi
uvicorn
pydantic
requests
supabase
joblib
scikit-learn
python-dotenv
numpy
